Moved EditorWorkerController to Interop/Connection
